Seperated the Login page with a Login form and Signup form
