+++ frontend/src/pages/Login/login.css
@@ -10,7 +10,7 @@ body {
 }
 
 /* Container */
-.container {
+.login-container {
     width: 100%;
     max-width: 400px;
     background: #fff;
